# ApiDocment.yaml
openapi: "3.0.0"
info:
  title: DXContestSupportSystem API(仮名)
  description: DXコンテストに関する活動においてサポートを行うシステムに関連するAPI
  version: 0.2.0 beta
security:
  - api_key: []
servers:
  - url: http://api.<domain>/v1
    description: 本番環境API
  - url: http://dev.<domain>/v1
    description: 開発環境API
paths:
...
  /room/{userId}:
    get:
      tags:
        - roomInOut
      summary: ユーザーの在室情報を取得
      parameters:
        - name: userId
          in: path
          description: ユーザーid
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 在室情報レスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomInOutResponse"
    patch:
      tags:
        - roomInOut
      summary: ユーザーの在室情報を更新(状態反転)
      parameters:
        - name: userId
          in: path
          description: ユーザーid
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 在室情報レスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomInOutResponse"
    post:
      tags:
        - roomInOut
      summary: ユーザーの在室情報を更新(状態指定)
      parameters:
        - name: userId
          in: path
          description: ユーザーid
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: ['in','out']
              
      responses:
        '200':
          description: 在室情報レスポンス
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomInOutResponse"
...
  /room/history/{userId}:
    get:
      tags:
        - roomInOut
      summary: ユーザーの入退室履歴を取得
      parameters:
        - name: userId
          in: path
          description: 学生Id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 入退室データ配列
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomInOut'
components:
  schemas:
...
    RoomInOut:
      type: object
      required:
        - id
        - userId
        - enterTime
      properties:
        id:
          type: integer
        userId:
          type: integer
        enterTime:
          type: string
          format: date
        leaveTime:
          type: string
          format: date
...
    RoomInOutResponse:
      type: object
      required:
        - userId
        - status
      properties:
        userId:
          type: string
        status:
          type: string
          enum: ['in','out']
        enterTime:
          type: string
          format: Date
...